# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация",
        "title": "ВИМБ API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/adv-messages": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список роликов, созданных в указанный период времени. Список отфильтрован по правам НВ.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочники"
                ],
                "summary": "Возвращает список роликов, созданных в указанный период времени. Список отфильтрован по правам НВ.",
                "operationId": "routes-get-adv-messages",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerGetAdvMessagesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/adv-messages/load": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание задач, на загрузку роликов, за выбранный период.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочники"
                ],
                "summary": "Создание задач, на загрузку роликов.",
                "operationId": "routes-load-adv-messages",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AdvMessagesLoadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
        "/api/v1/adv-messages/query": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Динамический запрос на загрузку сохраненных данных. Логические операторы: eq ==, ne !=, gt \u003e, lt \u003c, ge \u003e=, le \u003c=, in in, isnil is nil.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочники"
                ],
                "summary": "Загрузку сохраненных роликов.",
                "operationId": "routes-query-adv-messages",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AdvMessageQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommonResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
...
        "/api/v1/mediaplan": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый медиаплан.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Медиапланы"
                ],
                "summary": "Создает новый медиаплан.",
                "operationId": "routes-add-mediaplans",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerAddMPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает доступные медиапланы. Минимальный кванта периода начитки - 1 месяц (с первого по последнее число).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Медиапланы"
                ],
                "summary": "Возвращает доступные медиапланы. Минимальный кванта периода начитки - 1 месяц (с первого по последнее число).",
                "operationId": "routes-get-mediaplans",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerGetMPLansRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
...
        "/api/v1/spot": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Постановка спота в блок.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Постановка спота в блок.",
                "operationId": "routes-add-spot",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerAddSpotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список спотов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Возвращает список спотов.",
                "operationId": "routes-get-spots",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerGetSpotsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление спота в блоке.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Споты"
                ],
                "summary": "Удаление спота в блоке.",
                "operationId": "routes-delete-spot",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerDeleteSpotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StreamResponse"
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AdvMessageQuery": {
            "type": "object",
            "properties": {
                "AdtID": {
                    "type": "object",
                    "properties": {
                        "eq": {
                            "type": "integer",
                            "example": 700061957
                        }
                    }
                },
                "BrandID": {
                    "type": "object",
                    "properties": {
                        "eq": {
                            "type": "integer",
                            "example": 44362
                        }
                    }
                }
            }
        },
        "models.AdvMessagesLoadRequest": {
            "type": "object",
            "properties": {
                "Advertisers": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "AdvertisingMessageIDs": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Aspects": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "CreationDateEnd": {
                    "type": "string",
                    "example": "20191201"
                },
                "CreationDateStart": {
                    "type": "string",
                    "example": "20190301"
                },
                "FillMaterialTags": {
                    "type": "string",
                    "example": "true"
                }
            }
        },
...
        "models.CommonResponse": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.StreamResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "request": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerAddMPlanRequest": {
            "type": "object",
            "properties": {
                "BrandID": {
                    "type": "string"
                },
                "DateFrom": {
                    "type": "string"
                },
                "DateTo": {
                    "type": "string"
                },
                "MplCnlID": {
                    "type": "string"
                },
                "MplName": {
                    "type": "string"
                },
                "MultiSpotsInBlock": {
                    "type": "string"
                },
                "OrdID": {
                    "type": "string"
                }
            }
        },
        "models.SwaggerAddSpotRequest": {
            "type": "object",
            "properties": {
                "AuctionBidValue": {
                    "type": "string"
                },
                "BlockID": {
                    "type": "string"
                },
                "FilmID": {
                    "type": "string"
                },
                "FixedPosition": {
                    "type": "string"
                },
                "Position": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerDeleteSpotRequest": {
            "type": "object",
            "properties": {
                "SpotID": {
                    "type": "string"
                }
            }
        },
        "models.SwaggerGetAdvMessagesRequest": {
            "type": "object",
            "properties": {
                "Advertisers": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "AdvertisingMessageIDs": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Aspects": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "CreationDateEnd": {
                    "type": "string"
                },
                "CreationDateStart": {
                    "type": "string"
                },
                "FillMaterialTags": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerGetMPLansRequest": {
            "type": "object",
            "properties": {
                "AdtList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "AdtID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "ChannelList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Cnl": {
                                "type": "string"
                            }
                        }
                    }
                },
                "EndMonth": {
                    "type": "string"
                },
                "IncludeEmpty": {
                    "type": "string"
                },
                "SellingDirectionID": {
                    "type": "string"
                },
                "StartMonth": {
                    "type": "string"
                }
            }
        },
...
        "models.SwaggerGetSpotsRequest": {
            "type": "object",
            "properties": {
                "AdtList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "AdtID": {
                                "type": "string"
                            }
                        }
                    }
                },
                "ChannelList": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Cnl": {
                                "type": "string"
                            },
                            "Main": {
                                "type": "string"
                            }
                        }
                    }
                },
                "EndDate": {
                    "type": "string"
                },
                "InclOrdBlocks": {
                    "type": "string"
                },
                "SellingDirectionID": {
                    "type": "string"
                },
                "StartDate": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}